Move donate button to burger menu
And remove pull request section from donation dialog.

We've an issue upstream here as Tony is struggling to gather
the necessary funds to develop this project. This was addressed
partially by supporting his efforts on Patreon.

Having a donation button right there in the toolbar and
personal content (i.e. "my own use") in the donation dialog
breaks integration and makes the OS look amateurish though.

We might need to help with the development if the financial
help isn't enough, but we can't let the integration suffer.